+++ c/frame/replace.cc
@@ -1,17 +1,23 @@
 //------------------------------------------------------------------------------
 // Copyright 2018 H2O.ai
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+// Permission is hereby granted, free of charge, to any person obtaining a
+// copy of this software and associated documentation files (the "Software"),
+// to deal in the Software without restriction, including without limitation
+// the rights to use, copy, modify, merge, publish, distribute, sublicense,
+// and/or sell copies of the Software, and to permit persons to whom the
+// Software is furnished to do so, subject to the following conditions:
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
 //
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+// IN THE SOFTWARE.
 //------------------------------------------------------------------------------
 #include "frame/py_frame.h"
 #include <type_traits>
@@ -423,8 +429,7 @@ void ReplaceAgent::process_int_column(size_t colidx) {
   int64_t col_min = col->min();
   int64_t col_max = col->max();
   bool col_has_nas = (col->countna() > 0);
-  // xmax_int will be -MAXINT iff the replacement list has just one element: NA
-  if (xmin_int == std::numeric_limits<int64_t>::max()) {
+  if (xmin_int > xmax_int) {  // iff replace_what = [NA]
     if (!col_has_nas) return;
   } else {
     if (col_min > xmax_int || col_max < xmin_int) return;
@@ -481,8 +486,7 @@ void ReplaceAgent::process_real_column(size_t colidx) {
   double col_min = static_cast<double>(col->min());
   double col_max = static_cast<double>(col->max());
   bool col_has_nas = (col->countna() > 0);
-  // xmax_real will be NA iff the replacement list has just one element: NA
-  if (xmin_real == std::numeric_limits<double>::max()) {
+  if (xmin_real > xmax_real) {  // only when replace_what is [NA]
     if (!col_has_nas) return;
   } else {
     if (col_min > xmax_real || col_max < xmin_real) return;